Adding PDF copy of Peterson Master's Report.
I've had this PDF laying around forever, but I no longer seem to have
the LaTeX sources I used to generate it so many years ago.  Although
I'm somewhat dubious about some of the stuff I wrote here, it would be
a shame to lose it forever... so I'm uploading it.  Also this will
give me a public URL to use for citations.